REFACTOR: importing without /**/*/index.js use
